cut in half the panel events class
